requirements.txt
boto==2.39.0
botocore==1.4.13
certifi==2016.2.28
cffi==1.5.2
cryptography==1.3.1
dj-database-url==0.4.1
Django==1.8
django-allauth==0.25.2
django-markdown==0.8.4
django-parsley==0.6
django-stdimage==2.3.2
django-stdimage-serializer==0.1.2
django-storages==1.4.1
djangorestframework==3.3.3
docutils==0.12
enum34==1.1.3
Flask==0.10.1
futures==3.0.5
gunicorn==19.4.5
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
Markdown==2.6.5
MarkupSafe==0.23
ndg-httpsclient==0.4.0
oauthlib==1.1.1
Pillow==3.1.1
progressbar2==3.6.2
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
PyJWT==1.4.0
pyOpenSSL==16.0.0
python-dateutil==2.5.3
python-openid==2.2.5
python-resize-image==1.1.10
python-social-auth==0.2.19
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
Werkzeug==0.11.8
